Add aim indicators
Adds a conical, straight and detached circular
(for thrown AOE weapons) aim indicator. Player
now has a child Weapon node which describes the
range of the weapon as well as its type of
aim indicator.

The joystick mappings are also reworked to use
GetActionStrength instead of manually listening
for the joypad axes.