Add primary attack, improve weapon class system
Weapon is now an abstract class, which all
weapons will subclass.

StraightProjectileWeapon represents a weapon
that shoots a straight projectile at constant
intervals. It also adjusts its aim indicator
accordingly to fit the width of its bullets.
Gun is one of such a projectile weapons, and
will be the primary weapon of the first
character.

For now, the left trigger or left-click is used
for the primary attack, and the left bumper or E is
used for the secondary attack.

AttackJoyAxis and related names are also renamed to
WeightedAttackDirection.