Add overheating gun, add heat bar in UI
Gun has been refactored to OverheatingGun

OverheatingGun now heats up per projectile
fired. When it is not firing, it will cool
down at a constant rate.

When the gun's heat level is at or above
its maximum (1), the gun will be overheated.
An overheated gun will not be able
to fire until all of the heat is dissipated.

A basic heat level bar is added to the UI to
display the current heat level of the weapon.