MapLoader: Basic implementation
Walls are red, floors are black, white pixels mean nothing.
For now, each wall pixel generates to one wall unit block.
Adjacent floor pixels are grouped together so that a combined
mesh can be generated to save resources.

Some optimization could be made to reduce the number of meshes for
walls, but I think it's not a big issue for now.