Add enemies, basic dash skill
The dash skill allows the player to dash
forwards in a certain direction. For now,
he will either dash directly forwards, or
snap onto an enemy closest to his forward
direction, in a conical area.

For now, skills are to be controlled with
the attack joystick. When the secondary
attack button (LB/L1/RMB) is pressed,
the aim indicator of the player will swap
to that of the skill. The skill is then
activated upon release of the secondary
attack button.

The dash skill will show a small indicator
(a simple cube for now) above the target
enemy, if there is one. Dashing into enemies
will cause them to have a small knockback
effect into the air, for now. However,
this is only a partial implementation of
the skill. Future updates will further
work on the effects of this skill.

StraightProjectileWeapon has two callbacks
added: OnProjectileFired() and
OnProjectileHit(Enemy). Implementations of
StraightProjectileWeapon may override these
callbacks to specify the damage dealt to
the enemy on a per-projectile-per-enemy basis,
as well as create any effects once the
projectile hits an enemy.

StraightProjectileWeapon also has a method
CanFire(), which may be overridden by
implementations to specify whether the gun
will shoot a projectile. If CanFire() returns
false, the weapon will not shoot a projectile,
even if the player is trying to fire.

Also, the AimIndicators' rotations logic
has been tweaked. Instead of rotating the
indicators according to the attack direction,
the weapons are rotated, and the indicators
(being their child nodes) are simply facing
forwards locally.