Lay groundwork for status effects
This adds IStatusHolder and Status.

IStatusHolder is an interface that any entity
that can hold status effects implements. It
contains helper methods to add status effects
of certain durations, and check for presence
of status effects. This can be used by
external entities like weapons to modify
behaviour accordingly, or for the status holders
themselves to modify their appearance.

Status represents a status effect, which can
override ApplyStatus(float delta) to apply its
effects on the status holder. For example,
a poison effect can be achieved by reducing
the status holder's health in this method.

This system is currently being used to achieve the
"mark" effect required on the Overheating Gun.

Currently, the Overheating Gun marks any
enemies with a 30% chance per projectile, for
3 seconds. Marked enemies can then be attacked
using the dash skill to achieve a deadly combo.
Aforementioned deadly combo is to be implemented
in a future update.