Finish first character basics
The first character's skill set is in place.

Dashing to a marked enemy will now deal
damage to that enemy, and half as much to
other enemies in the area of impact around
the player.

Successful dash attacks are also cool the
overheating gun by 40% of its max heat level.

Weapon has been renamed to AimableAttack,
so as to allow for skills and weapons to
subclass from it.