Dash: Check if ray intersection with the enemy exists
Because targetEnemy was set some time before OnAttackButtonReleased
is executed, the enemy could have died or moved away at this point.
In this case, the ray would not intersect with anything, causing the
dictionary to not have the "position" key and the game will crash.

```
Unhandled Exception:
System.Collections.Generic.KeyNotFoundException: The given key was not present in the dictionary.
  at (wrapper managed-to-native) Godot.Collections.Dictionary.godot_icall_Dictionary_GetValue(intptr,object)
  at Godot.Collections.Dictionary.get_Item (System.Object key) [0x00000] in /root/godot/modules/mono/glue/GodotSharp/GodotSharp/Core/Dictionary.cs:114
  at DashSkill.OnAttackButtonReleased () [0x001b9] in /Users/daniellimws/coma-dash/Scripts/Weapon/DashSkill.cs:220
  at AimableAttack.set_IsAttackButtonPressed (System.Boolean value) [0x0001d] in /Users/daniellimws/coma-dash/Scripts/Weapon/AimableAttack.cs:38
  at Player._PhysicsProcess (System.Single delta) [0x000cd] in /Users/daniellimws/coma-dash/Scripts/Player.cs:126
```