Optimise floor generation in MapLoader
Changed colors used in MapLoader.
- White: walls
- Red: doors
- Green: starting tile
- Others: rooms

Removed old maps and updated biggest test map
to use new colors. Floor renamed to Room, and
Door has been added. For now, Door generates
the same mesh as Floor, but in the future we
can look to adding an actual door above the
floor.