Enemy: Add player following and attack
NoobEnemy provides an example on how we can move towards the player and
rotate to face the player using the MoveToPlayer and RotateToPlayer
methods exposed by the base Enemy class.

These functions are left to the children class to call themselves so
that they have full control over their actions.

NoobEnemy also provides an implementation for charging towards the
player to attack it. When it hits the player, the player will be knocked
back by a factor of the NoobEnemy's velocity.

As the saying goes, united we stand. When all enemies charge towards the player, the player will fly backwards by a large amount.

The NoobEnemy mesh is changed to a CSGBox so that the rotation can be
seen.

Additionally, add gravity to player.