AngstyEnemy: Enemy that charges and knocks back the player
AngstyEnemy provides an example on how we can move towards the player and rotate to face the player using the MoveToPlayer and RotateToPlayer methods exposed by the base Enemy class.

These functions are left to the children class to call themselves so that they have full control over their actions.

AngstyEnemy also provides an implementation for charging towards the player to attack it. When it hits the player, the player will be knocked back by a factor of the AngstyEnemy's velocity.

As the saying goes, united we stand. When all enemies charge towards the player, the player will fly backwards by a large amount.

The AngstyEnemy mesh is changed to a CSGBox so that the rotation can be seen.

Additionally, add gravity to player otherwise the player might fly out
of the game when hit.

Still doesn't have path finding towards the player, because the AStar class needs us to initialize it with some information about the room. So will do that after room gen is done.