Remove unnecessary floor in map generation
Turns out in setting UseCollision = true for the wall meshes, the
collider for the floor is also generated. So the extra floor static body
is no longer needed. I'm not sure how reliable is this though, but I
think can use this for now.

Also make player collider slightly longer in height otherwise the aim
indicator will fall below the ground and cannot be seen.