Parse map image to generate level meshes (#128)
* Ambitious but noob attempt at parsing bitmap myself

* Map: Add minimap for testing

* MapLoader: Parse map to generate polygons for mesh creation

* dotnet format

* Remove and gitignore mono crash dumps

* Remove redundant code in polygon tracing and fix door mesh

* Enable collision for csg meshes and add green pixel to floor bitmap

* Room: Split out setting shader params into one method

* DashSkill: Change type of GetOverlappingBodies elements to Spatial

Docs said that GetOverlappingBodies returns an array of PhysicsBodys but
apparently the array also contains a CSGPolygon, which will cause the
program to crash due to an invalid cast error.

* dotnet format

* Taller walls

* Remove unnecessary floor in map generation

Turns out in setting UseCollision = true for the wall meshes, the
collider for the floor is also generated. So the extra floor static body
is no longer needed. I'm not sure how reliable is this though, but I
think can use this for now.

Also make player collider slightly longer in height otherwise the aim
indicator will fall below the ground and cannot be seen.

* Code cleanup