MapLoader: Connect rooms with the doors next to them
Refactor the map parsing code to consider Wall/Room/Door as a Region.
Use a 2D map to store the type of region for each pixel. For each pixel,
if it is a door, check if any of its neighbouring pixels are a door, and
if so, connect the door with the room.

Room class exposes public methods to open or close doors that are
connected to it.