Add MeleeEnemy, HP regeneration
MeleeEnemy is an Enemy with a melee range.
When the player is within the range, it will
initiate the attack sequence.

1. The enemy charges the attack for a duration
   given by preAttackPauseDuration.
2. The enemy tries to attack the player
   immediately after the delay, dealing damage
   if the player is within range.
3. The enemy is then "paused" for another duration
   given by postAttackPauseDuration. This is to
   allow the player to execute melee hits more
   effectively.

Health regeneration is implemented in HealthEntity.
HealthEntities with a non-zero RegenerationRate will
begin to regenerate their health after not being
damaged for a duration given by RegenerationDelay, at
a rate specified by RegenerationRate.

A second progress bar is shown above the health bar,
indicating the time left until health regeneration
begins.