Make wall on bottom side of room transparent when player is in the room
* Create a `TransparentWall` class that allows toggling the wall's material
* Similar to connecting rooms with doors, connect rooms with
`TransparentWall` so that the room can toggle the wall's material
depending on whether the player is in the room
* This is made possible through `MapLoader.MarkTransparentWalls(...)` which creates a
new bitmap for the transparent walls by searching for walls located at the bottom
side of a room

Currently just proof of concept. Will try things like tweening the
transparency to make the effect look nicer.